requirements.txt
Package                 Version
----------------------- ------------------
absl-py                 2.2.2
accelerate              1.6.0
annotated-types         0.7.0
av                      14.3.0
certifi                 2025.1.31
charset-normalizer      3.4.1
click                   8.1.8
cloudpickle             3.1.1
contourpy               1.3.2
cycler                  0.12.1
docker-pycreds          0.4.0
docopt                  0.6.2
Farama-Notifications    0.0.4
filelock                3.16.1
fonttools               4.57.0
fsspec                  2024.10.0
ftfy                    6.3.1
gitdb                   4.0.12
GitPython               3.1.44
grpcio                  1.71.0
gymnasium               1.1.1
huggingface-hub         0.30.2
idna                    3.10
iniconfig               2.1.0
Jinja2                  3.1.4
kiwisolver              1.4.8
langcodes               3.5.0
language_data           1.3.0
llvmlite                0.44.0
locate                  1.1.1
marisa-trie             1.2.1
Markdown                3.8
MarkupSafe              2.1.5
matplotlib              3.10.1
mpmath                  1.3.0
msgpack                 1.1.0
networkx                3.4.2
num2words               0.5.14
numba                   0.61.2
numpy                   2.2.4
packaging               24.2
pandas                  2.2.3
pillow                  11.2.1
pip                     23.3.1
platformdirs            4.3.7
pluggy                  1.5.0
protobuf                5.29.4
psutil                  7.0.0
pydantic                2.11.3
pydantic_core           2.33.1
pygame-ce               2.5.3
pyparsing               3.2.3
pytest                  8.3.5
python-dateutil         2.9.0.post0
pytz                    2025.2
PyYAML                  6.0.2
qwen-vl-utils           0.0.10
regex                   2024.11.6
requests                2.32.3
safetensors             0.5.3
sb3_contrib             2.6.0
sentry-sdk              2.26.1
setproctitle            1.3.5
setuptools              68.2.2
six                     1.17.0
smmap                   5.0.2
stable_baselines3       2.6.0
sympy                   1.13.3
tensorboard             2.19.0
tensorboard-data-server 0.7.2
tokenizers              0.21.1
torch                   2.8.0.dev20250417
torchaudio              2.6.0.dev20250417
torchvision             0.22.0.dev20250417
tqdm                    4.67.1
transformers            4.50.0.dev0
typing_extensions       4.13.2
typing-inspection       0.4.0
tzdata                  2025.2
urllib3                 2.4.0
wandb                   0.19.9
wcwidth                 0.2.13
Werkzeug                3.1.3
wordfreq                3.1.1
